requirements.txt
flask==2.0.3 
boto3
python-dotenv
uwsgi
sentry-sdk[flask]
datadog-api-client


